refactor(nx-cmake): utils
