+++ packages/nx-cmake/src/index.ts
@@ -1,7 +1,7 @@
 import { type NxPluginV2 } from '@nx/devkit';
 import { PLUGIN_NAME } from './config/pluginName';
-import { createNodes } from './utils/graphUtils/createNodes/createNodes';
-import { createDependencies } from './utils/graphUtils/createDependencies/createDependencies';
+import { createDependencies } from './utils/createDependencies/createDependencies';
+import { createNodes } from './utils/createNodes/createNodes';
 
 const nxPlugin: NxPluginV2 = {
     name: PLUGIN_NAME,
