+++ packages/nx-cmake/src/utils/createNodesFunction/createNodesFunction.spec.ts
