+++ packages/nx-cmake/src/utils/getProjectName/getProjectName.spec.ts
@@ -1,5 +1,5 @@
 import { getProjectName } from './getProjectName';
-import { CProjectType } from '../../../models/types';
+import { CProjectType } from '../../models/types';
 
 describe('getProjectName', () => {
     it('should return the project name for an App project', () => {
